requirements.txt
langchain==0.1.5
langchain-huggingface==0.0.2
streamlit==1.29.0
python-dotenv==1.0.0